feat: portfolio-aware action plan — frontend + LLM injection + cost-basis sanitization + SearXNG dual-provider
Phase B frontend (originally staged):
- ActionPlanTable.tsx renders 持仓操作计划 cards with expandable 三维 rationale,
  direction emojis (⬆️⬇️🛑🎯), and contiguous priorities
- ReportSummary integrates the new section between overview and strategy
- TypeScript types include ActionPlanItem / CoreConclusion / DashboardSection

Agent path parity:
- _analyze_with_agent now threads self.portfolio_context_block into initial_context
- agent executor _build_user_message injects the portfolio block + action-plan
  instruction so the agent LLM also emits action_plan_items
- AGENT_SYSTEM_PROMPT + LEGACY example schemas include action_plan_items
- agent prompts drop "30字以内" cap on one_sentence to match non-agent path
- shared build_action_plan_instruction helper keeps both prompt paths in sync

LLM-injection post-processor (analyzer._try_inject_action_plan_items):
- focused, cost-basis-aware LLM call mirrors _try_inject_zh_translations
- sanitizes mini-model output:
  * drops take_profit items at or below avg_cost × 1.005 (not a real profit)
  * reclassifies stop_loss items above avg_cost × 1.02 as direction=sell
    (defensive trim, not loss prevention)
  * renumbers surviving priorities contiguously 1..N after filter
- updated prompt explicitly states the cost-basis constraint with a 口诀

Synthesis fallback (portfolio_context_service.synthesize_action_plan_items):
- shared _direction_for_stop_loss respects cost basis (same threshold as LLM filter)
- take_profit anchor clamps to avg_cost × 1.02 (break-even-plus) when chart TP ≤ cost
- quant_signal only populated from real quant data (volume_ratio / chip / turnover);
  no longer reuses volume_meaning narrative
- supports fractional-share accounts (Trading 212) without rounding errors
- parses avg_cost / last_price from rendered portfolio_context_block

API response wiring (the bug that hid action plans for fractional-share users):
- ActionPlanItemSchema.shares: Optional[int] → Optional[float] (was truncating
  0.25 → 0, then renderer dropped via `if not shares: continue`)
- analysis.py task-poll DB-fallback path now passes dashboard so action_plan_items
  survives server restarts
- analysis.py sync analyze path now passes dashboard
- frontend ActionPlanTable: null-safety on triggerPrice / shares; pctOfEquity uses
  != null so legitimate 0% values render; fractional shares formatted to 4 decimals

Renderer polish (both notification.py + history_service.py in parallel):
- "调仓策略" replaces "建仓策略" when portfolio_match=="held"
- "关键点位速查" hint line under 作战计划 when action_plan_items present
- 作战计划 操作点位 table header uses new trigger_price_label ("触发价" / "Trigger Price")

Multi-agent orchestrator:
- one_sentence fallback uses analysis_summary as-is (was re-truncating to 200 chars
  after summary was already capped at 220)

Search backend:
- SearXNGSearchProvider self-hosted instances and public-instance pool register
  as TWO separate providers in the fallback chain (was mutually exclusive). Chain:
  Tavily → self-hosted SearXNG → SearXNG-public. Tavily quota exhaustion no longer
  breaks news search.

Tests: 134 passing across action_plan + portfolio + agent + SearXNG suites.
Includes: cost-basis filter regressions, priority renumbering, fractional-share
schema validation, dual SearXNG provider registration.

+++ api/v1/schemas/history.py
@@ -112,7 +112,11 @@ class ActionPlanItemSchema(BaseModel):
     trigger_price: Optional[float] = None
     trigger_condition: Optional[str] = None
     direction: Optional[str] = None  # buy | sell | stop_loss | take_profit
-    shares: Optional[int] = None
+    # shares MUST be float — fractional-share accounts (Trading 212, Robinhood) routinely
+    # hold positions like 0.7597 shares. Typing as int caused Pydantic to coerce 0.25 → 0,
+    # the renderer dropped them via `if not shares: continue`, and entire action plans
+    # disappeared for users holding fractional shares.
+    shares: Optional[float] = None
     pct_of_position: Optional[float] = None  # null when not held
     pct_of_equity: Optional[float] = None
     technical_basis: Optional[str] = None
